save layouts and window positions across kwin/script resets
This keeps windows mostly where they were whenever kwin or bismuth
is restarted. Only the window sizes aren't saved.

Layout state is now shared per each <monitor, group> pairing.
That means if you have group 1 on monitor 1 of desktop 1 and you
switch to desktop 2 and recall group 1 on monitor 1, it will
share the same layout state, whereas if instead you recall it to
monitor 2 (on any desktop) it will instead have its own layout state.
This seems desirable for example if monitor 1 is portrait yet
monitor 2 is landscape, as the same layout is unlikely to work well
on both.